St. Anne allowing community to see all reservations
In DashboardPresenter.php St. Anne's community decided that they wanted everyone to see all upcoming reservations and not just their own.  We generally don't have anything to hide in this community.  A better solution would be to make a configuration available in the app settings instead of using the custom code in this commit.